% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_report.R
\name{archive_report}
\alias{archive_report}
\title{Archive a rendered report}
\usage{
archive_report(
  file = "report.html",
  overwrite = FALSE,
  quiet = getOption("lbstproj.quiet", FALSE)
)
}
\arguments{
\item{file}{\emph{Character}. The name of the report file to archive.

\emph{Default}: \code{"report.html"}, since this is the default file name in the report
generation process.}

\item{overwrite}{\emph{Logical}. Indicate whether to overwrite an existing archive file.

\emph{Default}: \code{FALSE}, an error is raised if the archive file already exists.}

\item{quiet}{\emph{Logical}. If \code{TRUE}, suppress informational messages. Important
messages (e.g. directory creation or errors) are still shown.

\emph{Default}: \code{FALSE} unless the global option \code{lbstproj.quiet} is set otherwise.}
}
\description{
Moves a rendered report from the \verb{report/} directory to the
\verb{results/reports/} directory and renames it to include the current date and
project version. This function is typically used after rendering a report
with \code{\link[=run_report]{run_report()}}, to save a copy of the rendered output.
}
