% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{build_report_name}
\alias{build_report_name}
\title{Build a report file name with date, version, and optional counter and ext}
\usage{
build_report_name(extension, counter = NULL)
}
\arguments{
\item{extension}{The file extension for the report (e.g., "html", "pdf").}

\item{counter}{An optional counter to append to the file name for uniqueness.
If \code{NULL}, no counter is added.}
}
\description{
Build a report file name with date, version, and optional counter and ext
}
