% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_path.R
\name{check_can_overwrite}
\alias{check_can_overwrite}
\title{Check if a file can be overwritten}
\usage{
check_can_overwrite(path, overwrite, what = "File")
}
\arguments{
\item{path}{\emph{Character}. Path of the file to check}

\item{overwrite}{\emph{Logical}. Whether the file can be overwritten or not.}

\item{what}{\emph{Character}. The type of file being checked. This is only used to customize the error message if the file can't be overwritten.

\emph{Default}: \code{"File"}. A generic description}
}
\value{
Invisibly return the path of the file
}
\description{
Check if a file can be overwritten
}
\keyword{internal}
