% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_file.R
\name{create_data}
\alias{create_data}
\title{Create a data R file}
\usage{
create_data(name, open = TRUE, quiet = getOption("lbstproj.quiet", FALSE), ...)
}
\arguments{
\item{name}{\emph{Character}. File name (with or without \code{.R}). Only letters,
numbers, and hyphens are allowed.}

\item{open}{\emph{Logical}. Open the file after creation.}

\item{quiet}{\emph{Logical}. If \code{TRUE}, suppress informational messages.

\emph{Default}: \code{TRUE} unless the global option \code{lbstproj.quiet} is set otherwise.}

\item{...}{Additional fields passed as data to the file template.}
}
\value{
Invisibly returns the path of the file created.
}
\description{
Shortcut for \code{create_file(type = "data", ...)}. Creates an R
script in \verb{R/data/} using the data template.
}
\examples{
if (FALSE) {
create_data("import-adsl")
# > v Import-adsl file created at 'R/data/import-adsl.R'
}

}
