% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_description}
\alias{create_description}
\title{Create a DESCRIPTION file for the project}
\usage{
create_description(
  path,
  title,
  client,
  department,
  author,
  version,
  quiet = FALSE
)
}
\arguments{
\item{path}{Path where the project should be created (default is current
directory).}

\item{title}{Title of the project.}

\item{client}{Name of the client.}

\item{department}{Department code or name.}

\item{author}{Author name (first and last).}

\item{version}{Project version (default is \verb{1.0.0}).}
}
\description{
Create a DESCRIPTION file for the project
}
\details{
This function does not check arguments for validity. It assumes
that they are checked in the \code{create_project()} function.
}
\keyword{internal}
