% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chunk.R
\name{create_figure_chunk}
\alias{create_figure_chunk}
\title{Create a Rmd code chunk to insert a figure}
\usage{
create_figure_chunk(id, print = TRUE)
}
\arguments{
\item{id}{The id of the figure as specified in the table of tables (TOT)}

\item{print}{Logical, whether to print the chunk to the console (default: TRUE)}
}
\description{
Create a Rmd code chunk to insert a figure
}
