% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_readme}
\alias{create_readme}
\title{Create a README.md file for the project}
\usage{
create_readme(path, quiet = FALSE, overwrite = FALSE)
}
\arguments{
\item{path}{Path to the project directory.}

\item{quiet}{If \code{TRUE}, suppresses messages (default is \code{FALSE}).}

\item{overwrite}{If \code{TRUE}, overwrites existing README.md file}
}
\description{
Create a basic README.md file in the root of the project directory. It
contains the project title, author, and a brief overview of the directory
structure.
If the file already exists, it will not be overwritten.
}
\details{
This function does not check arguments for validity. It assumes
that they are checked in the \code{create_project()} function.
}
\keyword{internal}
