% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_file.R
\name{create_scripts}
\alias{create_scripts}
\alias{create_figure}
\alias{create_table}
\alias{create_function}
\alias{create_data}
\alias{create_model}
\title{Create a R script pre-filled with boilerplate code}
\usage{
create_figure(name, open = rlang::is_interactive(), ...)

create_table(name, open = rlang::is_interactive(), ...)

create_function(name, open = rlang::is_interactive(), ...)

create_data(name, open = rlang::is_interactive(), ...)

create_model(name, open = rlang::is_interactive(), ...)
}
\arguments{
\item{name}{Name of the script (without file extension).}

\item{open}{Whether to open the new file in the editor. Defaults to \code{TRUE} if
in an interactive session (i.e. in RStudio and other IDEs).}

\item{...}{Additional arguments passed to the template rendering function.
See details.}

\item{print}{Whether to print (default: \code{TRUE}) a success message to the
console. You can set a default print behavior for all functions in this
family by setting the global option \code{use.print} to \code{TRUE} or \code{FALSE}.}
}
\description{
A family of helpers that write a minimal R script into a project
subfolder under \verb{R/}: figures, tables, functions, and data.
}
\details{
These functions are called for their side effects: they create a new
script, make parent directories if needed, and (optionally) open the file.
They do not return a value.

When creating a new \emph{table} or a new \emph{figure} script, you can pass the additional
argument \code{id}. It will be used to retrieve caption and label information from
the table of tables of the project. This allow dynamic changes to captions
and labels in a single place (i.e. the table of tables).
}
\section{Functions}{
\itemize{
\item \code{create_figure()}: Create a figure script in \verb{R/figures/}.

\item \code{create_table()}: Create a table script in \verb{R/tables/}.

\item \code{create_function()}: Create a function script in \verb{R/functions/}.

\item \code{create_data()}: Create a data script in \verb{R/data/}.

\item \code{create_model()}: Create a model script in \verb{R/models/}.

}}
\examples{
\dontrun{
create_figure("eda_hist01")
create_table("summary_endpoints", open = FALSE)
create_function("strings_utils")
}
}
