% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_tot}
\alias{create_tot}
\title{Create a Empty Table of Tables (TOT) Excel file in the project}
\usage{
create_tot(quiet = FALSE)
}
\description{
\code{create_tot()} copies a template TOT Excel file from the \code{lbstproj} package
to the \verb{data/tot/} directory of your project. The example TOT file contains
an example row for guidance. It should be deleted before use.
}
\details{
If a "data/tot/" directory does not exist, it will be created.
If a TOT file already exists in the target directory, it will be overwritten.
}
\keyword{internal}
