% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{ensure_dir_exists}
\alias{ensure_dir_exists}
\title{Ensure directory exists}
\usage{
ensure_dir_exists(dir, confirm = TRUE, print = getOption("use.print", TRUE))
}
\arguments{
\item{dir}{Character. The path of the directory to check/create.}

\item{confirm}{Logical. If \code{TRUE}, prompts the user for confirmation before
creating the directory. If \code{FALSE}, creates the directory without
confirmation.}

\item{print}{Logical. If \code{TRUE}, prints a message to the CLI about the
creation status of the directory. Defaults to the global option \code{use.print}
(which defaults to \code{TRUE}).}
}
\value{
Invisibly returns the relative path of the directory if it exists or
}
\description{
Checks if a directory exists and creates it if it doesn't. If the session is
non-interactive, an error is thrown instead of prompting the user. If the
directory already exists, the function silently ends.
}
\keyword{internal}
