% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_path.R
\name{ensure_file_does_not_exist}
\alias{ensure_file_does_not_exist}
\title{Ensure a file does not exist}
\usage{
ensure_file_does_not_exist(file)
}
\arguments{
\item{file}{\emph{Character}. File path. Must be relative to active project
and not absolute.}
}
\value{
Invisibly returns the normalized path to the file, relative to
the active project.
}
\description{
Verify if the path to a file exists and throws an error if it the case. This is used
purely for validation purposes.
}
\keyword{internal}
