% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tot.R
\name{import_tot}
\alias{import_tot}
\title{Import the Table of Tables (TOT)}
\usage{
import_tot(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical. If \code{TRUE}, suppress informational messages.

\emph{Default}: \code{FALSE} unless the global option \code{lbstproj.quiet} is set otherwise. The default option can be changed using \code{options(lbstproj.quiet = TRUE)}}
}
\value{
Invisibly returns the TOT data as a tibble.
}
\description{
Reads the TOT Excel file from \code{data/tot/table_of_tables.xlsx}, validates its
structure, and saves it as an RDS file in \code{data/tot/tot.rds}.
}
