% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_gt_project}
\alias{is_gt_project}
\alias{is_flextable_project}
\title{Test whether the active project uses a given table engine}
\usage{
is_gt_project()

is_flextable_project()
}
\value{
\emph{Logical}. \code{TRUE} when the project's table engine matches.
}
\description{
Test whether the active project uses a given table engine
}
\keyword{internal}
