% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbstproj-package.R
\docType{package}
\name{lbstproj-package}
\alias{lbstproj}
\alias{lbstproj-package}
\title{lbstproj: Tools To Create And Use Standardized R Projects For The L-Biostat Group.}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provides functions to easily start a new statistical analysis project in R. It creates a standardized folder structure and files, to allow for easier collaboration and reproducible research. It also provides functions to easily load and preprocess data, generate and export figures and tables, and create reports in various formats.
}
\author{
\strong{Maintainer}: Alexandre Bohyn \email{alexandre.bohyn@kuleuven.be} (\href{https://orcid.org/0000-0001-9776-7467}{ORCID})

}
\keyword{internal}
