% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_func}
\alias{load_func}
\alias{load_all_functions}
\alias{load_function}
\title{Load functions from the \code{R/functions} directory}
\usage{
load_all_functions()

load_function(name)
}
\arguments{
\item{name}{The name of the specific function file to load. It must be
relative to the \code{R/functions} directory and can include or omit the \code{.R}
extension.}
}
\description{
These helper functions load one or all functions from the \code{R/functions}
directory.
}
\section{Functions}{
\itemize{
\item \code{load_all_functions()}: Load all function files.

\item \code{load_function()}: Load a single function file.

}}
\examples{
\dontrun{
load_all_functions()
load_function("data_cleaning") # Works
load_function("data_cleaning.R") # Also works
load_function("R/functions/data_cleaning.R") # Fails
}

}
