% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tot.R
\name{load_tot}
\alias{load_tot}
\title{Load the Table of Tables (TOT)}
\usage{
load_tot(refresh = TRUE)
}
\arguments{
\item{refresh}{Logical; if \code{TRUE}, re-imports the TOT from the Excel file
using \code{import_tot()}. Default is \code{TRUE} because the TOT may be updated
frequently.}
}
\value{
A tibble containing the TOT data with the following columns:
\itemize{
\item \code{id}: Unique identifier for each item.
\item \code{type}: Type of the item (either "table" or "figure").
\item \code{name}: Name of the item.
\item \code{caption}: Caption of the item.
}
}
\description{
Simple wrapper function to load the TOT RDS file.
}
\examples{
\dontrun{
 tot <- load_tot()
}
}
