% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{move_rename_report}
\alias{move_rename_report}
\title{Moves a report to a different location and rename it uniquely}
\usage{
move_rename_report(old_path, overwrite = FALSE)
}
\arguments{
\item{old_path}{The current path of the report file to be moved (with extension in the
filename)}

\item{overwrite}{Logical; if \code{TRUE}, overwrite existing report files in the \verb{results/reports/} directory.    If \code{FALSE}, a unique name will be generated to avoid overwriting, by appending a counter to the filename (e.g. \verb{_001}, \verb{_002}, etc.). Default is \code{FALSE}.}
}
\value{
The path to the moved and renamed report file.
}
\description{
Move the report file generated by \code{run_report()} from the \verb{reports/} directory to the
\verb{results/reports/} directory, renaming it to include:
\itemize{
\item the current date
\item the project version (from the DESCRIPTION file)
\item a unique identifier if a file with the same name already exists
This function helps to archive reports systematically.
}
}
