% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_data}
\alias{save_data}
\title{Save a processed dataset to the project}
\usage{
save_data(
  data,
  name,
  overwrite = TRUE,
  quiet = getOption("lbstproj.quiet", FALSE)
)
}
\arguments{
\item{data}{\emph{A \code{data.frame} object}. The data to be saved.}

\item{name}{\emph{Character}. File name (without extension). Must only contain
letters, numbers, and hyphens}

\item{overwrite}{\emph{Logical}. If \code{TRUE}, overwrite existing file with the same
name. If \code{FALSE}, an error is thrown if a file with the same name already
exists.

\emph{Default}: \code{TRUE}}

\item{quiet}{\emph{Logical}. If \code{TRUE}, suppress informational message for file
saving.

\emph{Default}: \code{TRUE} unless the global option \code{lbstproj.quiet} is set
otherwise.}
}
\description{
This function saves a \code{data.frame} as an \code{.rds} file in the \verb{data/processed/}
folder of the project. If the folder does not exists, it will be created. The
file name is validated to ensure that it only contains letters, numbers, and
hyphens.
}
