% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_object}
\alias{save_object}
\title{Internal function to save an R object to the appropriate directory}
\usage{
save_object(type, object, name, overwrite, print = NULL, ...)
}
\description{
Internal function to save an R object to the appropriate directory
}
\keyword{internal}
