% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{save_report}
\alias{save_report}
\title{Save a report file}
\usage{
save_report(file = "report.qmd", overwrite = FALSE)
}
\arguments{
\item{file}{The name of the report file to save. Default is \code{"report.qmd"}.}

\item{overwrite}{Logical indicating whether to overwrite an existing file.
Defaults to \code{FALSE}, so an error is raised if the target file already exists.}
}
\description{
Rename a report file located in the \verb{report/} directory to a unique name
including the current date and project version. This function is used when
the current version of a report needs to be preserved before making further
changes (e.g. write methodology, document new analyses, ...), because
\code{\link[=create_report]{create_report()}} will automatically overwrite existing report file.
}
\examples{
\dontrun{
# Creates a file named like "report_interim_2024_06_15_v1.0.0.qmd" in the
# same folder
save_report("report_interim.qmd")
}

}
