% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_data_raw}
\alias{use_data_raw}
\alias{use_table}
\alias{use_figure}
\alias{use_function}
\title{Create a skeleton script}
\usage{
use_data_raw(name, open = rlang::is_interactive(), overwrite = FALSE)

use_table(name, open = rlang::is_interactive(), overwrite = FALSE)

use_figure(name, open = rlang::is_interactive(), overwrite = FALSE)

use_function(name, open = rlang::is_interactive(), overwrite = FALSE)
}
\arguments{
\item{name}{Name of the dataset/table/figure/function. This will be sanitized
and used as the script file name.}

\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}

\item{overwrite}{By default, all the \verb{use_*} functions will not overwrite
existing files. Set to \code{TRUE} if you really want to do so.}
}
\description{
These functions set up a directory (if needed) and create a new R script from
a template. The script serves as a starting point for preparing data,
generating tables, creating figures, or writing helper functions in your
project.
}
\details{
Depending on the function you use, the script will be placed in one of the
following directories:
\itemize{
\item \code{use_data_raw()}: \verb{R/data/}
\item \code{use_table()}: \verb{R/tables/}
\item \code{use_figure()}: \verb{R/figures/}
\item \code{use_function()}: \verb{R/functions/}
}
}
