% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_file}
\alias{use_file}
\title{Internal function to create a R-script file from a template based on type}
\usage{
use_file(type, name, overwrite, open, print = NULL, ...)
}
\description{
Internal function to create a R-script file from a template based on type
}
\keyword{internal}
