% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_path.R
\name{validate_file_name}
\alias{validate_file_name}
\title{Validate the name of a script}
\usage{
validate_file_name(name)
}
\arguments{
\item{name}{\emph{Character}. The name of the script to validate.}
}
\value{
\emph{Character}. The validated name of the script without any extension.
}
\description{
Ensures that the name of a R script only contains valid characters (letters, numbers, underscores, and hyphens) and remove any potential extension.
}
\keyword{internal}
