% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{archive_report}
\alias{archive_report}
\title{Archive a rendered report}
\usage{
archive_report(file = "report.html", overwrite = FALSE)
}
\arguments{
\item{file}{The name of the report file to archive. Default is \code{"report.html"}.}

\item{overwrite}{Logical indicating whether to overwrite an existing archive file.
Defaults to \code{FALSE}, so an error is raised if the archive file already exists.}
}
\description{
Moves a rendered report from the \verb{report/} directory to the
\verb{results/reports/} directory and renames it to include the current date and
project version. This function is typically used after rendering a report
with \code{\link[=run_report]{run_report()}}, to save a copy of the rendered output.
}
