% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{build_report_name}
\alias{build_report_name}
\title{Build a report file name with date, version, and optional counter and extension}
\usage{
build_report_name(extension, counter = NULL)
}
\description{
Build a report file name with date, version, and optional counter and extension
}
