% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chunk.R
\name{create_chunk}
\alias{create_chunk}
\title{Standardized way to create a Rmd code chunk to insert a figure or a table}
\usage{
create_chunk(id, type, print = TRUE)
}
\arguments{
\item{id}{The id of the figure as specified in the table of tables (TOT)}

\item{print}{Logical, whether to print the chunk to the console (default: TRUE)}
}
\value{
The rendered Rmd chunk as a character string, invisibly
}
\description{
Standardized way to create a Rmd code chunk to insert a figure or a table
}
