% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chunk.R
\name{create_chunk}
\alias{create_chunk}
\title{Standardized way to create a Rmd code chunk to insert a figure or a table}
\usage{
create_chunk(id, type, copy = TRUE, print = TRUE, pad = FALSE)
}
\arguments{
\item{id}{The id of the figure as specified in the table of tables (TOT)}

\item{type}{The type of the entry, either "figure" or "table"}

\item{copy}{Logical, whether to copy the chunk to the clipboard (default: TRUE if in interactive mode)}

\item{print}{Logical, whether to print the chunk to the console (default: TRUE)}

\item{pad}{Logical, whether to pad the chunk with blank lines before and after and a pagebreak after (default: FALSE)}
}
\value{
The rendered Rmd chunk as a character string, invisibly
}
\description{
Standardized way to create a Rmd code chunk to insert a figure or a table
}
