% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_description}
\alias{create_description}
\title{Create a DESCRIPTION file for the project}
\usage{
create_description(
  path,
  title,
  client,
  department,
  author,
  version,
  quiet = FALSE
)
}
\description{
Create a DESCRIPTION file for the project
}
\details{
This functions does not check arguments for validity. It assumes
that they are checked in the \code{create_project()} function.
}
