% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{create_file}
\alias{create_file}
\title{Internal function to create a R-script file from a template based on type}
\usage{
create_file(type, name, open, print = NULL, ...)
}
\description{
Internal function to create a R-script file from a template based on type
}
\keyword{internal}
