% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_programs.R
\name{create_programs}
\alias{create_programs}
\title{Create all R scripts listed in the table of tables (TOT)}
\usage{
create_programs()
}
\description{
This function loads the TOT and creates all the R scripts for figures and
tables listed there in their respective directories in the \code{R} folder. Any
existing scripts with the same names will NOT be overwritten.
}
