% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create a New Project}
\usage{
create_project(
  path = ".",
  title = NULL,
  client = NULL,
  department = NULL,
  author = NULL,
  version = "1.0.0",
  open = rlang::is_interactive(),
  force = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{path}{Path where the project should be created (default is current
directory).}

\item{title}{Title of the project.}

\item{client}{Name of the client.}

\item{department}{Department code or name.}

\item{author}{Author name (first and last).}

\item{version}{Project version (default is \verb{1.0.0}).}

\item{open}{If \code{TRUE}, opens the new project in RStudio (default is \code{TRUE} if
in interactive session).}

\item{force}{If \code{TRUE}, skips the confirmation prompt before creating the
project (default is \code{FALSE}).}

\item{quiet}{If \code{TRUE}, suppresses informational messages during project
creation (default is \code{FALSE}).}
}
\value{
Invisibly returns the active project path.
}
\description{
\code{create_project()} initializes a new project with a standardized structure at
the specified path. It prompts the user for essential metadata if not
provided. The function creates the necessary directories, a \code{DESCRIPTION}
file, and an table of tables (TOT) Excel file. If the project is successfully
created, it opens the new project in RStudio.
}
\details{
The following directory structure is created inside the project:

\if{html}{\out{<div class="sourceCode">}}\preformatted{├── data
│   ├── figures
│   ├── processed
│   ├── raw
│   ├── tables
│   └── tot
├── docs
│   └── meetings
├── results
│   ├── figures
│   ├── tables
│   └── reports
├── R
│   ├── data
│   ├── figures
│   ├── functions
│   └── tables
└── report
    └── utils
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
create_project(
  path = ".", # uses current directory
  title = "Example Project",
  client = "Client Name",
  department = "DEP",
  author = "Jane Doe",
  version = "0.1.0"
)
}
}
