% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_readme}
\alias{create_readme}
\title{Create a README.md file for the project}
\usage{
create_readme(path, title, author, quiet = FALSE)
}
\description{
Create a basic README.md file in the root of the project directory. It
contains the project title, author, and a brief overview of the directory
structure.
If the file already exists, it will not be overwritten.
}
\details{
This functions does not check arguments for validity. It assumes
that they are checked in the \code{create_project()} function.
}
