% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Create a quarto report based on the table of tables (TOT)}
\usage{
create_report()
}
\description{
Creates a quarto report by using a html-report template and including all
elements found in the TOT in the order they appear. All captions and labels
are taken from the TOT, so any changes to those will be reflected in the
report when it is rendered.
}
\details{
The report is created in \code{report/report.qmd}. If a report already exists, it
will be overwritten.
}
