% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_structure}
\alias{create_structure}
\title{Create the standard project structure}
\usage{
create_structure(path, quiet = FALSE)
}
\description{
\code{create_structure()} creates the standard folder structure for the project.
Any folders that already exist will be skipped.
}
\details{
This functions does not check arguments for validity. It assumes
that they are checked in the \code{create_project()} function.
}
