% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_path.R
\name{ensure_dir_exists}
\alias{ensure_dir_exists}
\title{Ensure a directory exists}
\usage{
ensure_dir_exists(path, create = TRUE)
}
\arguments{
\item{path}{\emph{Character}. Directory path. Must be relative to active project
and not absolute.}

\item{create}{\emph{Logical}. If \code{FALSE}, throws an error if the directory does
not exist. If \code{TRUE} creates it.

\emph{Default}: \code{TRUE}}
}
\value{
Invisibly returns the normalized path to the directory, relative to
the active project.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Create a directory (recursively) if it does not exist already or throw an
error if needed. Paths must be relative to the active project, not absolute.
}
\keyword{internal}
