% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{lbstproj-options}
\alias{lbstproj-options}
\title{Global options for the lbstproj package}
\description{
Global options for the lbstproj package
}
\section{Global options for printing messages}{

\itemize{
\item \code{options(save.print = TRUE)} (default): print messages when saving figures and tables.
\item \code{options(export.print = TRUE)} (default): print messages when exporting figures and tables.
\item \code{options(use.print = TRUE)} (default): print messages when creating a new R script.
}
}

