% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{lbstproj-options}
\alias{lbstproj-options}
\title{Globel options for the lbstproj package}
\description{
Globel options for the lbstproj package
}
\section{Global options for printing messages}{

\itemize{
\item \code{options(save_info = TRUE)} (default): print messages when saving figures and tables.
\item \code{options(export_info = TRUE)} (default): print messages when exporting figures and tables.
}
}

