% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all.R
\name{run_all}
\alias{run_all}
\alias{run_all_figures}
\alias{run_all_tables}
\title{Run All R Scripts in ToT}
\usage{
run_all_figures(skip = NULL)

run_all_tables(skip = NULL)
}
\arguments{
\item{skip}{Optional vector of non-negative integers corresponding to the
\code{id} column in the ToT. Any scripts with these IDs will be skipped.}
}
\description{
Run all R scripts listed in the Table of Tables (ToT) for a
specified type (figure or table).
}
\section{Functions}{
\itemize{
\item \code{run_all_figures()}: run all scripts of type "figure"

\item \code{run_all_tables()}: run all scripts of type "table"

}}
