% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{run_report}
\alias{run_report}
\title{Render the report}
\usage{
run_report()
}
\value{
The path to the moved and renamed report file.
}
\description{
Render the Rmd report located at \code{report/report.Rmd} to the \verb{report/} directory.
Also move and rename the generated report to the \verb{results/reports/} directory.
}
