% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{run_report}
\alias{run_report}
\title{Render a report}
\usage{
run_report(file = "report.qmd", ...)
}
\arguments{
\item{file}{The name of the report file to render. Default is \code{"report.qmd"}, the
default name created by \code{\link[=create_report]{create_report()}}.}

\item{...}{Additional arguments passed to the rendering function. For R Markdown
files, these are passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}. For Quarto files, these are
passed to \code{\link[quarto:quarto_render]{quarto::quarto_render()}}.}
}
\description{
Renders a report file located in the \verb{report/} directory. The report can be
either an R Markdown (\code{.Rmd}) or Quarto (\code{.qmd}) file. The rendered output
is saved in the same \verb{report/} directory.
}
\details{
All report files are render to a file with the same name. If such a file already
exists, it will be overwritten without warning. If you want to keep a copy of the
rendered report, use the \code{\link[=archive_report]{archive_report()}} function after rendering to move it to
the \verb{results/reports/} directory with a unique name.
}
\examples{
# Default usage
run_report()
# Rendering a specific R Markdown file to pdf
run_report(file = "pdf_report.Rmd", output_format = "pdf_document")
}
