% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_data}
\alias{save_data}
\alias{save_table}
\alias{save_figure}
\title{Save an R object}
\usage{
save_data(x, name, overwrite = FALSE, quiet = FALSE, ...)

save_table(x, name, overwrite = FALSE, quiet = FALSE, ...)

save_figure(x, name, overwrite = FALSE, quiet = FALSE, ...)
}
\arguments{
\item{x}{The object (data, table, figure) to be saved.}

\item{name}{A name for the object. Should only contain letters, numbers, and underscores.}

\item{overwrite}{By default, all the \verb{save_*} functions will not overwrite
existing files. Set to \code{TRUE} if you really want to do so.}

\item{quiet}{If \code{TRUE}, suppresses messages about the saving process. Defaults to \code{FALSE}.}
}
\description{
These functions save R objects to the appropriate directories within the project
structure, ensuring that they exist. All objects are saved as \code{.rds} files to be easily
loaded later.
}
\details{
Each function saves the object to a specific directory:
\itemize{
\item \code{save_data()} to \verb{data/processed/}.
\item \code{save_table()} to \verb{data/tables/}.
\item \code{save_figure()} to \verb{data/figures/}.
}
}
\examples{
\dontrun{
# Example of saving a data frame
df <- data.frame(x = 1:10, y = rnorm(10))
save_data(df, "my_dataset")
}
}
